supervisor_v2/pyproject.toml
[project]
name = "robot-buddy-supervisor-v2"
version = "0.1.0"
description = "Raspberry Pi 5 supervisor for robot-buddy (v2 — process-isolated workers)"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
    "fastapi>=0.110",
    "httpx>=0.27",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
]

[project.optional-dependencies]
rpi = [
    "picamera2>=0.3.17",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["supervisor_v2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::pytest.PytestConfigWarning"]
